Fix broken unit tests
A mock object was declared in a file which contained unit tests.
When something is imported from a file containing tests, the tests
in that file are run. This means that when a test failed in that
file, jest would report that the same test failed multiple times.
The solution was to move the mock object to a separate file from
the tests.
